created an auth service to allow a user to login and provide data on the current user.
created a user model.

updated the navbar to import in the auth service and added in an ngIf to show either a login link or the current user. currently routing to the events route as soon as login is clicked without validation. Cancel also lead to to the events route.