created an upvote component and its style sheet, as well as a voter service to handle adding, deleting or whether the user has voted.
created a toggle vote function on the session list component to either add or delete votes and update the sort by.  also created a user has boted function to check if the currently logged in user has already voted for a given session.

added in a votes column to the session list component view.